+++ core/src/main/java/com/expediagroup/streamplatform/streamregistry/core/predicate/VersionPredicateFactory.java
@@ -18,6 +18,7 @@ package com.expediagroup.streamplatform.streamregistry.core.predicate;
 import java.util.Comparator;
 
 import com.google.common.collect.ImmutableListMultimap;
+
 import org.springframework.stereotype.Component;
 
 import com.expediagroup.streamplatform.streamregistry.model.Stream;