+++ repository-kafka/src/main/java/com/expediagroup/streamplatform/streamregistry/repository/kafka/Config.java
@@ -26,6 +26,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeoutException;
+
 import javax.annotation.PostConstruct;
 
 import lombok.AccessLevel;