+++ repository-kafka/src/main/java/com/expediagroup/streamplatform/streamregistry/repository/kafka/StoreProducer.java
@@ -23,6 +23,7 @@ import lombok.RequiredArgsConstructor;
 import io.confluent.kafka.serializers.KafkaAvroSerializer;
 import io.confluent.kafka.serializers.KafkaAvroSerializerConfig;
 import io.confluent.kafka.serializers.subject.TopicRecordNameStrategy;
+
 import org.apache.avro.specific.SpecificRecord;
 import org.apache.kafka.clients.producer.KafkaProducer;
 import org.apache.kafka.clients.producer.Producer;