+++ repository-kafka/src/main/java/com/expediagroup/streamplatform/streamregistry/repository/kafka/StoreView.java
@@ -32,6 +32,7 @@ import lombok.extern.slf4j.Slf4j;
 import io.confluent.kafka.serializers.AbstractKafkaAvroSerDeConfig;
 import io.confluent.kafka.serializers.subject.TopicRecordNameStrategy;
 import io.confluent.kafka.streams.serdes.avro.SpecificAvroSerde;
+
 import org.apache.avro.specific.SpecificRecord;
 import org.apache.kafka.common.utils.Bytes;
 import org.apache.kafka.streams.KafkaStreams;